WIP: *Actually* make reference counting work.
Godot wants to be in control of the memory and we ignore that at our own
peril. Thus we introduce Ref[T] which wraps a given T: RefCounted (not
enforced as of yet). Furthermore, all wrapped classes are now `ptr object`
and Ref[T] is in charge of making sure it's freed at the appropriate
time.

TODO:
 - Passing Ref[T]s to and from function parameters needs to use ref_set_object
   and ref_get_object to synchronize with Ref<T> on the Godot side.
   (apparently, to be tested)

 - Manually managed classes for now leak memory as there is no more
   destructor. Introduce Owned[T] which forbids =copy.