Bindgen: Drop jsony for std/json
jsony was a good choice when the bindgen was running in nimvm as part of
the actual build process because std/json would not even finish parsing
without exploding the instruction budgets, but since even that was too
slow for development use, bindgen is now a compiled step and this is no
longer necessary.

Also removes a compiler warning and I hate compiler warnings.